+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # main
 
+* Added cop `Grape/HelpersIncludeModule` ([#1](https://github.com/petalmd/rubocop-petal/pull/1))
+
 # v0.1.0
 
 * First version